Add Database#supports_{index_parsing,foreign_key_parsing,table_listing,view_listing}?
Currently, the only way to check for support is to call and
rescue NotImplemented.  Exceptions for flow control is not really
a great idea, so it's better to just check for support.  In
Sequel 4, the methods will just not be defined if they aren't
implemented, so these will become respond_to? checks.