Deprecate passing an options hash to Database#dataset or Dataset.new
Instead of requiring that an options hash be passed, require
that no options be passed.  This forces you to have to clone
to modify datasets.  Looking at the existing code, almost nothing
passed an argument hash to these methods.

Instead of cloning a hash with deprecated methods, just extend a plain
hash with the deprecated methods, as that appears to be better for
performance.