Deprecate Model#set_values
Unfortunately, this only deprecates the overriding of the method,
not the calling of it.  There isn't a way to deprecate the calling
of the method that doesn't result in a ton of of spurious
deprecation warnings or breaks existing code.

The reasoning behind this is that set_values was never designed
to be used by external code.  It was only public because Model.call
calls it.  In 4.0, Model.call will just set the hash value directly.
The two other users of it are manual refresh calls and then automatic
refreshes from the database after insert.  Those two cases are
currently handled the same way, but they will be split in 4.0,
since they result in undesireable behavior (clearing caches in the
automatic refresh case that don't need to be cleared.