Deprecate automatic typecasting of hash and array to string for string columns
Almost everything in ruby responds to to_s, so that is loose by
default.  Since automatic conversion from hash/array seems like
an unlikely desire, and can easily be triggered by an attacker,
Sequel 4 will start raising exceptions for it.

looser_typecasting's purpose is to make typecasting loose, so
it's a natural fit to move the previous default string typecast
to that extension.