Make constraint_validations plugin reflect validations by column
This adds reflection information when using the
constraint_validations plugin, similar to the support in the
validation_class_methods plugin.  The two are not compatible due
to differences in the underlying API.

To implement this easily, the private constraint_validation_array
now takes two arguments, the second being the hash of reflections.