Deprecate parsing columns for the same table name in multiple schemas on jdbc
The basic idea is if you try to parse the schema for an unqualified
table, you can get columns from tables with the same name in multiple
schemas.  The shared postgres adapter had code similar to this before
it was changed to allow disambiguation.

I looked into disambiguation options on JDBC, but the JDBC metadata
method that is supposed to allow it (ResultMetaData getSchemaName)
doesn't appear to work on either Microsoft SQL Server or DB2, and
does not appear reliable enough.

I debated just raising an error instead of adding a deprecate
warning, since the current code is likely to cause an error, but it's
possible with models that someone is getting schema information
and not using it, and just breaking their code would be impolite.