Emulate offsets on Microsoft Access using reverse orders and total counts (Fixes #661)
This is terrible for performance, but I think that's better than
not working at all.  It's still has some corner cases, such as
requiring that you select the columns you are ordering on, but
hopefully they won't be problematic.

Note that this should work for Microsoft SQL Server 2000 as well,
though we don't support that.  For people using Sequel with
Microsoft SQL Server 2000, they can manually extend their datasets
with the offset emulation module.