Validate against explicit nil values in NOT NULL columns with default values in the auto_validations plugin
If an explicit nil value is present in the values, then Sequel
will use an explicit NULL value when inserting the row, which
will cause a DatabaseError even if the column has a default
value.  So use a non_null :allow_missing=>true validation for
such columns to catch such errors.  Explicitly validate the
primary key in the same way, since it operates similarly but
in certain databases the schema parsing shows a nil default
value.