Switch association_proxies plugin to use a hash of options for proxy decision
Previously, it just took three arguments: the method name, method
arguments, and method block.  However, sometimes you might want
to change where to send the method based on the instance or
reflection, or on how the association was created.  This revised
API allows you to do this.

While here, fix a bug in the previous code where the block
passed to the association method was not passed to the association
code.