Remove default graph_each, and each calling graph_each
Make the graph_each plugin handle this.  Since each no longer
calls graph_each, make the model associations code override each
for the eager graph checking.  This is only a temporary situation,
as eager_graph in the future will call naked on the dataset so that
each does not need to be overridden.