Remove Dataset.introspect_all_columns from columns_introspection.
While here, also remove columns_without_introspection, and just use
super in the overridden columns.

As the specs used introspect_all_columns if an environment
variable was used, which them to using the extension with all
databases.  However, to work around issues with the mock dataset
columns accepting multiple arguments, specifically include the
module in the mock dataset class, so that extending a mock
dataset with columns_introspection doesn't override the setting
of column information via arguments.