Model#save no longer calls #_refresh internally, deprecate Model#set_values
This separates the three types of set_values calls:

1) When loading objects from the database (Model.call)

2) When manually refreshing objects to discard existing changes
   (Model#refresh).

3) When automatically refreshing objects to get the default values
   for model columns (Model#save).

For some plugins, this causes three duplicated code paths, since
those plugins want the same behavior for each of these paths.

However, other plugins wanted different behavior for the three
paths, but since they used set_values, ended up with the same
behavior for each.

This changes the associations, composition, serialization plugins
to not clear cached values in case 3).  Any cached values should
not change in an after_save refresh, so clearing the caches just
caused more work in the database, and possible issues in the
application.

This makes the dirty plugin not clear cached values twice.

This adds the _refresh_set_values and _save_set_values private
methods to replace cases 2) and 3).  Case 1) doesn't have a method,
as Model.call just sets the instance variable directly in the
instance.