ungraphed on an eager_graph dataset now resets the original row_proc
Previously, Dataset#each was overridden, and eager_graph was handled
specially to avoid calling the row_proc.  Since that's just the same
as the naked behavior, it's better just to call naked.

However, plenty of users abuse eager_graph by calling ungraphed
on the result, to get a dataset with the joins done, but not wanting
a graphed result.  In that case, for backwards compatibility, restore
the previous row proc so that they get a model object instead of
an empty hash.

Note that this is not backwards compatible if you call naked between
eager_graph and ungraphed.  That's probably uncommon enough that it
does not warrant a deprecation warning.