Refactor set-restricted based mass assignment methods
This changes set_restricted and update_restricted to accept a
single type argument instead of separate only and except
arguments.  For set/update/new/create, a value of :default is
used.  For set_all/update_all, a value of :all is used.  For
set_only/update_only, an array of allowed columns is used.

To make the blacklist_security implementation easier,
an uncached .get_setter_methods is added, and blacklist_security
just overrides that.  To handle the set_except/update_except
methods, an array subclass is used that is handled specially in
the overridden Model#setter_methods.

This simplifies quite a bit of the internal code.  The only
backwards incompatibility is in the private methods.