Use prepared statements when the association has :conditions in the prepared_statements_associations plugin
This tweaks the previous commit to compare the unbound variables
to the association variables, and bind any that are missing.  This
should allow safe prepared statement usage for associations with
conditions.

Note that this only works because conditions should not be able
to contain instance-specific code.  You can use a proc for conditions,
but then it is a virtual row block and doesn't have access to the
dataset.  Both the block form and the :dataset option allow for
instance-specific code, so prepared statements should not be used
for either of them.