Make dataset string literalization that requires databaset connection use dataset's chosen server
Previously, the transparent master/slave behavior offered by
Sequel fell down in string literalization, since it always used the
default server for dataset literalization.

This uses the dataset's server if one is specified.  Note that this
is still not enough to get transparent behavior unless you have a
server set for all datasets.