Raise exception for infinite and NaN floats on MySQL (Fixes #677)
MySQL doesn't support direct input of these, and on 5.1.24+
automatically translates any computation yielding them to NULL.
The previous behavior sent bad SQL to the server, this just has
Sequel raise an exception instead of sending a query.

It's possible in the future that the behavior could be customized,
but as it's unlikely someone will need it, there is no reason to
do so at this point.