Support creating/refreshing/dropping materialized views on PostgreSQL
Materialized views were added to PostgreSQL in 9.3, and this adds
support for them.  create_view and drop_view get a :materialized
option, and a refresh_view method was added for refreshing the
materialized views.

While here, I noticed the IF EXISTS syntax wasn't supported by
drop_view on PostgreSQL, so add that while here.