Support LATERAL queries via Dataset#lateral
To use LATERAL queries, you tag the dataset you want to use with
Dataset#lateral, and when the subquery is literalized, LATERAL
is prepended to it:

  DB.from(:a, @dataset.from(:b).lateral).sql
  # SELECT * FROM a, LATERAL (SELECT * FROM b) AS t1"

LATERAL is supported on Postgres 9.3rc1+, HSQLDB, and DB2. It can
be emulated on Microsoft SQL Server using the
mssql_emulate_lateral_with_apply extension, which uses CROSS/OUTER
APPLY for the emulation.