Rename INTEGRATION_DB and MODEL_DB to DB in specs
This makes the specs use Sequel's general convention of using DB
for the Sequel::Database object.

The MODEL_DB usage dates back to before 1.0.  The INTEGRATION_DB
was my own creation, modeled after the database specific constants
in the adapters.  The reason the adapter tests used separate
constants originally was mostly likely that they were run in the
same process (looking at the sequel_core Rakefile for 1.0). Since
it's been many years since tests for separate adapters were run
in the same process, there was no need to keep them separate, so
the adapter constants were made the same as the integration
constant in the earlier commit.  This commit just makes them use
the shorter and more familiar DB constant.