Correctly raise error when using an invalid virtual row block function call
Because VirtualRow is a BasicObject subclass, and raise is an Object
method, not a keyword, raise inside a VirtualRow doesn't raise an
exception, it treats it as calling the raise database function, you
need to use Kernel.raise to raise an exception.  Not caught previously
by the specs as the specs also checked for literalization, and an
exception is raised later when trying to literalize the exception
class.