Allow CHECK constraints to have options by using an options hash as the constraint name
Previously, the API did not allow for check constraints to have
options, since all arguments to the constraint method were used
for the constraint itself.  This allows you to pass an options
hash as the constraint name, allowing you to use constraint
options.

This is most useful to allow NOT VALID to be used when adding
CHECK constraints to existing tables on PostgreSQL.