Add from_block Database extension, for having Database#from block affect FROM instead of WHERE
I wish I realized this was an issue a few months ago, so I could
have made this the default in Sequel 4.  But this change is too
large to go into a minor release, so I won't be able to make it
the default until Sequel 5.  In the meantime, this extension
makes it easy to get the desireable behavior.