Add pg_static_cache_updater Database extension for listening for changes to tables and updating static_cache caches automatically
This makes it possible to use the static_cache plugin for models
where the underlying table can change at runtime.  Assuming you
meet the criteria for using the extension, it should now be
possible to use the static_cache plugin for any model whose
dataset is small and not frequently updated.