Remove hooks updating the cache by default from the static_cache plugin
While this initially seemed like an easy way to keep the cache in
sync, it would only work if you were using a single process to make
changes.  As soon as you scale to multiple processes, it wouldn't
work.  So it doesn't make sense to even offer it as a default.

The only safe way to do this is to have the cache invalidation be
initiated by the database, as the pg_static_cache_updater handles
it.