Add implementations of Model#dup and Model#clone
Previously, these didn't have special implementations, which meant
that Object#dup and Object#clone were used.  Unfortunately, that
meant that the underlying data structures such as values were
shared, which is not desired.

The implementation of #dup follows the implementation of #freeze,
if #freeze would freeze an internal data structure, then #dup
will dup it.

Unfortunately, Model#clone cannot be implemented in the same
way.  Because Object#clone copies the frozen flag before
returning, you cannot modified the cloned object returned to
duplicate the internal data structures.  To work around this
issue, Model#clone is basically the same as #dup, except that
if the receiver is frozen, it will freeze the dup before
returning it.  This has the negative consequence that other
things that Object#clone does (such as copying singleton
classes and extended modules) do not happen when using
Model#clone.