Add implementations of Dataset#freeze and Dataset#dup
Previously, Dataset#freeze broke the dataset, since you couldn't
clone it afterward.  Inside of using ruby's frozen flag, store
the frozen flag in opts.  Datasets are generally treated as frozen
anyway, with just the mutation methods modifying the receivers
instead of returning cloned objects, so add a manual check to
the mutation methods to check for frozen status.

Make dup the same as clone, but have it clear the frozen flag.

In order for frozen? to work, add a separate implementation that
just checks for the frozen flag in opts.