Better typecast on load references in pg_* extension RDoc
For pg_array and pg_hstore, in some cases you need to use the
regular typecast_on_load extension, since the jdbc/postgres
can return the values as Array or Hash instances instead of
Strings, and pg_typecast_on_load only works with strings.

For the other ones, just update the text to be consistent.