Add error_sql Database extension, allowing DatabaseError#sql to return SQL query that caused underlying exception
This is mostly helpful when the underlying database doesn't use
good error messages, but I suspect there could be other uses for
it.

Unfortunately, the implementation is pretty hacky, and there
are some corner cases.  This really can't be helped, as the
DatabaseError exceptions are not created until after the SQL
used is out of scope.