Add table_select plugin for selecting table.* instead of * for model datasets
Several people have noted issues when joining model datasets
to other tables, where the returned columns have values from the
other table's columns overriding the model's columns. The reason
this occurs is the default selection for all Sequel datasets is *,
since no other default selection makes sense (a dataset by
default does not even have a table).  Because Sequel returns
record sets as hashes, and the adapters generally add values to
result set hashes in column order (with later ones with the same
name overriding previous ones with the same name), if you joined
a model to another table that has columns with the same name (a
common usage), you can end up with result sets where the column
value is the value of the column with the same name in the other
table, instead of the model's table.

ActiveRecord uses a default of SELECT table.* FROM table, allowing
you to join your model to another table and not have the other
model columns be put in your result set.  This plugin makes Sequel
more similar to ActiveRecord in that area, which should ease the
transition for some users.

Note that it's probably better to use explicit selections any time
you join tables.  For one, when joining a model dataset to another
table manually, the chances that you need either:

1) all columns in the model table and none in the other table; or
2) all columns in both tables

are small.  In general you need some columns from the model table
and some columns from the other table.