Add support for foreign key parsing on Oracle
I was testing the generic jdbc foreign key support, and it didn't
work on Oracle, due to a fairly literal interpretation of how
the JDBC methods work (only including foreign keys referencing
primary keys, not other unique keys).

This dataset is based on the JDBC query, but cut down and modified
to also deal with foreign keys reference unique keys.

Oracle also doesn't like duplicate unique indexes, and there
doesn't appear to be a reason to use them in the spec.