Fix a couple issues with association cloning
This fixes two issues with association cloning.  The first issue
is that :eager_block was not cloned from the source dataset if
it wasn't set specifically via an option (the default is to use
the block passed to the method).

The second issue is that cloning an association that was a clone
of another association didn't have the block from the original
association.