Make filter by associations support handle associations with :conditions or block
This fixes a documented limitation with the filter by associations
support, where it previously did not handle associations that
were not entirely specified by the keys.  This allows the filter
by associations to work with associations with :conditions and
most type of blocks, by adding an extra subselect that tests
that the object(s) used as the filter by associations value also
meet the criteria for the association.