Make offset emulation on Oracle work when using columns that can't be ordered (Fixes #724)
This changes Oracle from using the generic ROW_NUMBER window
function emulation to using a specific ROWNUM emulation. The
generic ROW_NUMBER window function can add orders on all columns,
and Oracle doesn't allow you to order on some column types.

Fix the oracle spec added to test this, since it tests for
ordered output without specifying an order, and that's not safe,
as without an order specified, the database is free to return
rows in any order.