Refactor jdbc adapter conversion proc handling
This abstracts the column type name retrieval to a method that
subadapters can override to get different column information.
It also only calls this method if the subadapter actually needs
this information (most don't), which should be better for
performance.

Additionally, it refactors the code to reduce the number of
branches and method calls.

For the sqlanywhere subadapter, do the smallint to bool conversion
correctly, looking first at the dataset setting and then at the
database setting.  Previously, both had to be true or it wouldn't
do the conversion. Also, rename the boolean type translator method
to sqla_boolean, since it is specific to the sqlanywhere subadapter.

Also fix the sqlanywhere convert_smallint_to_bool dataset setting,
which should default to the database setting, not the global default.