Improve PostgreSQL type support in the jdbc/postgresql adapter
This makes the PostgreSQL type handling in the jdbc/postgresql
much closer to the postgres adapter type handling, using the
same conversion procs in most cases.  This means that in most
cases, you no longer need to use the pg_typecast_on_load or
typecast_on_load plugins to handle these types in models, since
the datasets return the correct results automatically.

One of the other improvements is this allows you to return types
that were previously not handled at all before in the
jdbc/postgresql adapter, such as arrays of other custom types.
Previously trying to return such types resulted in a JDBC not
implemented error.