Support OFFSET/FETCH for ordered datasets on Microsoft SQL Server 2012
Microsoft SQL Server 2012 adds native support for offsets, but
support is limited to datasets that have orders, and using it
requires a different limit syntax than TOP.  While it has limitations,
most people using offsets are going to be using them with limits
and orders, so it handles most actual use cases.

This changes Sequel to use the new offset syntax on SQL Server 2012
for ordered queries, instead of emulating offsets with a subquery
using row_number.