Don't use a subselect when using offset for unordered datasets on MSSQL 2012+
This extends the previous commit to not use a subselect even in
the unordered case.  Since Sequel's offset emulation via row number
requires adding an order anyway on MSSQL, this should just eliminate
the subquery.

There is a slight difference, in that because the order is being
specifed in the ORDER BY clause instead of the over clause, we can
just order by the first column instead of having to get the column
names (which can cause problems if aliases are used).  This should
cause no problems in practice, since if the user did not specify
an order, the database is free to return results in an arbitrary
order.