Work with RSpec 3
This updates the specs to switch be_true and be_false with == true
and == false, since be_true and be_false are renamed in RSpec 3.

Avoid deprecations warnings on RSpec 3 by explicitly enabling only
the should syntax.  This has the benefit of not allowing the expect
syntax, which is good as that doesn't work on older versions of
rspec (and Sequel's specs should pass on RSpec 1).