Improve and Generalize the tinytds stored procedure support
Make the code generic using a dataset instead of the adapter
specific API.  As datasets return hashes with symbol keys,
the API has changed and now uses symbol keys in the returned
hash, as well as simpler key names.

Move the code to the mssql shared adapter so that all adapters
that connect to MSSQL can use it.  Unfortunately, do to bugs
in ruby-odbc, it does not work with the odbc adapter.

Modify the necessary modifications so that it runs correctly
on ruby 1.8.

Reduce possibilities for SQL injection in the output
variable name by quoting the name used in select, while using
a consistent name for the variable itself.