Add association_join to model datasets, for setting up joins based on associations
There have been many requests for this type of functionality in the
past, but I could never think of a good name for it.  I've given up
on a good name, so association_join it is.

Implementation is fairly simple, piggy-backing on the eager_graph
support, and just copying the resulting JOIN.  Only a couple new
features were needed:

1) Making the default eager_grapher procs respect a passed in
:eager_graph :join_type option and override the one used for the
association.

2) Making Dataset#graph not use a from_self in cases, as otherwise
using separate association_join calls doesn't work (since the
:graph/:eager_graph metadata is retained across calls).