Use an eager limit strategy by default for *_one associations with orders
If an order is specified for a *_one association, then the association
is probably not a true one-to-one association, and therefore could
probably benefit from an eager limit strategy.

Previously, this was only done automatically for *_one associations
if they had offsets.

With this change, the distinct on eager limit strategy can be used
by default.