Simplify queries used for filtering by associations with conditions
Previously, the strategy Sequel used for associations with conditions
was to start with the filter for the association without a condition,
and then add another filter for the conditions part.  However, the
filter for the conditions in general just repeated most of the code
for the filter for the association without the condition.  This was
especially bad for associations involving joins, as there would be
multiple subqueries instead of a single subquery.

This new code just drops the filter for the association without a
condition, fixing a few corner cases for many_*_many associations
where an IS NOT NULL condition was not added.

While here, add an association reflection helper method for
returning the dataset with the eager block condition already added.
This reorders some clauses, but as the specs would be changing anyway
due to removal of other clauses, it makes sense to do it at the same
time.