Use limit strategy to correctly handle filtering by limited associations
This applies the same type of strategy as eager loading limited
associations to filtering by limited associations.

If a limited association is used when filtering by associations,
Sequel automatically uses the more complex filter that would be
used if the association had conditions, but with an additional
subquery in that filter to restrict the results to only those
where the object meets the limit requirements for the association.

This change can cause a significant slowdown, but it is required
to get correct results, and my assumption is that anyone doing:

  Artist.where(:first_10_albums=>Album[1])

really wants the artists where that album is one of their first
10 albums, not artists where that album is their 11th album.