Add SQL::AliasedExpression#alias alias for #aliaz
alias is a keyword in ruby, so when I originally wrote the code
I didn't want to use it.  However, having method names that overlap
with keywords is fine as long as you call the methods with an
explicit receiver.  I think this makes the calling code easier to
read.