Avoid unnecessary parens in some generated SQL
This makes three separate but related changes to remove extra
parens.

When creating complex expressions with an arbitrary number of
arguments, strip off any NOOP expressions.  This will allow
things like AND(foo, NOOP(AND(bar, baz)) to combine into
AND(foo, bar, baz).

When using Sequel.& and related methods, in the single argument
case, return the argument directly if it is already of the
correct class.  If the single argument is not already of the
correct class, wrap it in a NOOP.