Add Dataset#paged_each :strategy=>:filter option for increased performance
This new strategy makes #paged_each not use offsets, but instead use
filters based on the last retreived row to implement the paging.
Using this new filtering approach is signficantly faster than using
the previous offset approach.  In simple testing:

                indexed     unindexed
100000 rows        7.3x          4.1x
200000 rows       13.4x          4.6x
300000 rows       19.9x          4.6x

Unfortunately, determining when this approach is safe to use is
difficult.  In SQL, identifiers appearing in the ORDER BY clause
can be aliased in the SELECT clause, and the ORDER BY clause can
refer to either the identifier or the alias, and they can be
ambiguous (SELECT a AS b, b AS a FROM c ORDER BY a, b), and I'm not
sure if all databases handle the rules per the SQL standard (it
would surprise me if so).  It's also possible to order on columns
not being selected, in which case you can't get the values for the
last retrieved row.  Also, if the value for any of the column values
for the last retrieved row are NULL, you can't use a filtering based
approach.

Basically, there are too many corner cases to turn this on by
default, which is why it requires manually specifying the option.

To handle nontrivial cases, when using the :strategy=>:filter option,
you can also provide a :filter_value option proc.  This proc takes
the last retrieved row and an array of order by expressions, and
should return an array of values in the last retrieved row related
to those expressions.  The default is to guess at the symbols used
and just look in the hash for them, but that can fail in some of the
cases mentioned above.