Use a cursor to implement paging for Dataset#paged_each in the postgres adapter
Using cursors to implement paging should always be faster than the
default strategies used by paged_each, and it also has fewer
restrictions (works on unordered datasets), so it should be the
default.

Currently, I've made it the only approach, since I can't think of
a reason you'd want to use one of the other strategies.  If for
some reason one of the other strategies makes sense, I'll be happy
to offer a way to use one of the default strategies.